+++ src/commands/mod.rs
@@ -1,2 +1,4 @@
 pub mod messages;
 pub mod roles;
+pub mod user;
+pub mod dev;
