+++ src/main.rs
@@ -1,8 +1,7 @@
 #![warn(clippy::str_to_string)]
 
-mod commands;
-mod commands_old; //legacy commands to be moved into the /commands/ dir
 mod utils;
+mod commands;
 
 use poise::serenity_prelude as serenity;
 use std::{
@@ -11,31 +10,11 @@ use std::{
     time::Duration,
 };
 
-// Types used by all command functions
-type Error = Box<dyn std::error::Error + Send + Sync>;
-type Context<'a> = poise::Context<'a, Data, Error>;
+use utils::{
+    base::Data,
+    handlers::on_error,
+};
 
-// Custom user data passed to all command functions
-pub struct Data {
-    votes: Mutex<HashMap<String, u32>>,
-}
-
-async fn on_error(error: poise::FrameworkError<'_, Data, Error>) {
-    // This is our custom error handler
-    // They are many errors that can occur, so we only handle the ones we want to customize
-    // and forward the rest to the default handler
-    match error {
-        poise::FrameworkError::Setup { error, .. } => panic!("Failed to start bot: {:?}", error),
-        poise::FrameworkError::Command { error, ctx, .. } => {
-            println!("Error in command `{}`: {:?}", ctx.command().name, error,);
-        }
-        error => {
-            if let Err(e) = poise::builtins::on_error(error).await {
-                println!("Error while handling error: {}", e)
-            }
-        }
-    }
-}
 
 #[tokio::main]
 async fn main() {
@@ -44,7 +23,14 @@ async fn main() {
     // FrameworkOptions contains all of poise's configuration option in one struct
     // Every option can be omitted to use its default value
     let options = poise::FrameworkOptions {
-        commands: vec![commands_old::help(), commands_old::ping(), commands_old::vote(), commands_old::getvotes(), commands_old::poll()],
+        commands: vec![
+            commands::messages::help(), 
+            commands::messages::vote(), 
+            commands::messages::getvotes(), 
+            commands::messages::poll(),
+            commands::dev::ping(),
+            commands::user::check(),
+        ],
         prefix_options: poise::PrefixFrameworkOptions {
             prefix: Some("!".into()),
             edit_tracker: Some(Arc::new(poise::EditTracker::for_timespan(
